Switched an unsigned to size_t. They aren't the same everywhere.

+++ cpp/unix.c
@@ -93,7 +93,7 @@ setup(int argc, char **argv)
 /* memmove is defined here because some vendors don't provide it at
    all and others do a terrible job (like calling malloc) */
 void *
-memmove(void *dp, const void *sp, unsigned int n)
+memmove(void *dp, const void *sp, size_t n)
 {
 	unsigned char *cdp, *csp;
